refactor(node/sync): extract peer-selection helpers into sync::peers (Phase 1 of #2313)
First slice of the `SyncManager` decomposition. Extracts the
peer-discovery + anchor-priority partition logic that
`perform_interval_sync` previously inlined into ~150 lines, plus
the `partition_peers_anchor_first` free function that was sitting
at the bottom of `manager/mod.rs`.

New file `crates/node/src/sync/peers.rs` owns:

- `discover_mesh_peers_with_namespace_fallback(...)` — async
  helper that retries `mesh_peers` on the context topic up to
  `max_retries` rounds, falls back to the namespace-topic mesh
  if the context topic stays empty (the resolution closure is
  owned by the caller so this module doesn't depend on
  `ContextClient`), and bails with a diagnostic if both meshes
  are empty after the retry budget. Returns a `DiscoveryOutcome`
  carrying the peer list, which mesh produced them, attempts
  used, and elapsed time.

- `partition_peers_anchor_first(...)` — moved verbatim from
  `manager/mod.rs`. Already a free function taking
  `&dyn SyncStateAccess`; just relocated.

Tests live alongside the module and exercise both functions
against `MockSyncNetwork` + `MockSyncStateAccess` /
`crate::state::NodeState`. The partition tests previously in
`manager/tests.rs` move with the function.

`perform_interval_sync` now calls
`super::peers::discover_mesh_peers_with_namespace_fallback`
and unwraps the outcome, replacing the inline 90-line discovery
block. The namespace-resolution closure captures
`context_client.clone()` for the `get_context_group_id` +
`resolve_namespace` walk that has to stay in the manager
(can't move into `peers.rs` without that module depending on
`calimero-context`).

Net diff: `manager/mod.rs` shrinks by 136 lines;
`manager/tests.rs` shrinks by 196 lines; one new file of 465
lines (the bulk is tests).

Behaviour change: none.
