+++ crates/node/src/sync/manager/tests.rs
@@ -336,202 +336,6 @@ mod session_watchdog_tests {
 }
 
 // =========================================================================
-// `partition_peers_anchor_first` — anchor-preferred peer ordering
-// =========================================================================
-//
-// Contract: stable order within each (anchor / non-anchor) partition.
-// The caller pre-shuffles for randomness; this helper just hoists the
-// anchor partition to the front without reordering within it.
-//
-// These exercise the pure-function shape. The integration (looking up
-// the anchor identity set from the store and threading it through
-// `perform_interval_sync`) is covered by the `trusted_anchors_*`
-// helper-level tests in `calimero-context` and end-to-end by multi-node
-// e2e runs.
-
-use std::collections::BTreeSet;
-
-use calimero_primitives::identity::PublicKey;
-use dashmap::DashMap;
-use libp2p::PeerId;
-
-use super::partition_peers_anchor_first;
-
-/// Build a `NodeState` populated with the given `(peer, identities)`
-/// pairs in its `peer_identities` cache and nothing else, so the
-/// `SyncStateAccess` impl on `NodeState` can serve the partition
-/// tests without standing up the full node. The remaining trait
-/// methods are unreachable from `partition_peers_anchor_first` —
-/// it only calls `peer_identities` — so leaving them at their
-/// default (empty) state is safe.
-fn node_state_with_peer_identities(
-    entries: impl IntoIterator<Item = (PeerId, BTreeSet<PublicKey>)>,
-) -> crate::state::NodeState {
-    let node_state = crate::state::NodeState::new(false, crate::run::NodeMode::Standard);
-    for (peer, ids) in entries {
-        let _replaced = node_state.peer_identities.insert(peer, ids);
-    }
-    node_state
-}
-
-fn dummy_peer(n: u8) -> PeerId {
-    // Deterministic peer-id keyed by a single byte — only equality
-    // matters here, not the byte structure.
-    let seed = [n; 32];
-    let kp = libp2p::identity::Keypair::ed25519_from_bytes(seed).expect("valid seed");
-    PeerId::from_public_key(&kp.public())
-}
-
-fn dummy_pk(n: u8) -> PublicKey {
-    PublicKey::from([n; 32])
-}
-
-#[test]
-fn partition_empty_anchors_set_returns_zero() {
-    // No anchor set defined → no preference; every peer is non-anchor.
-    let mut peers = vec![dummy_peer(1), dummy_peer(2), dummy_peer(3)];
-    let node_state = node_state_with_peer_identities([]);
-    let anchors: BTreeSet<PublicKey> = BTreeSet::new();
-
-    let count = partition_peers_anchor_first(&mut peers, &node_state, &anchors);
-    assert_eq!(count, 0);
-}
-
-#[test]
-fn partition_empty_cache_no_anchors_found() {
-    // Anchor set non-empty but we've observed nothing → fall back to
-    // non-anchor; relative order preserved.
-    let mut peers = vec![dummy_peer(1), dummy_peer(2)];
-    let original = peers.clone();
-    let node_state = node_state_with_peer_identities([]);
-    let anchors: BTreeSet<PublicKey> = [dummy_pk(0xAA)].into_iter().collect();
-
-    let count = partition_peers_anchor_first(&mut peers, &node_state, &anchors);
-    assert_eq!(count, 0);
-    assert_eq!(peers, original);
-}
-
-#[test]
-fn partition_all_peers_are_anchors() {
-    let peer1 = dummy_peer(1);
-    let peer2 = dummy_peer(2);
-    let pk_admin = dummy_pk(0xAA);
-    let node_state = node_state_with_peer_identities([
-        (peer1, [pk_admin].into_iter().collect()),
-        (peer2, [pk_admin].into_iter().collect()),
-    ]);
-    let anchors: BTreeSet<PublicKey> = [pk_admin].into_iter().collect();
-
-    let mut peers = vec![peer1, peer2];
-    let count = partition_peers_anchor_first(&mut peers, &node_state, &anchors);
-    assert_eq!(count, 2);
-    assert_eq!(peers, vec![peer1, peer2]);
-}
-
-#[test]
-fn partition_mixed_anchor_and_non_anchor_preserves_relative_order() {
-    let anchor_a = dummy_peer(1);
-    let anchor_b = dummy_peer(2);
-    let plain_a = dummy_peer(3);
-    let plain_b = dummy_peer(4);
-    let pk_admin = dummy_pk(0xAA);
-
-    let node_state = node_state_with_peer_identities([
-        (anchor_a, [pk_admin].into_iter().collect()),
-        (anchor_b, [pk_admin].into_iter().collect()),
-    ]);
-
-    let anchors: BTreeSet<PublicKey> = [pk_admin].into_iter().collect();
-
-    // Pre-shuffled order interleaves the two partitions.
-    let mut peers = vec![plain_a, anchor_a, plain_b, anchor_b];
-    let count = partition_peers_anchor_first(&mut peers, &node_state, &anchors);
-    assert_eq!(count, 2);
-    // Anchors first (anchor_a before anchor_b, preserved from input),
-    // then non-anchors (plain_a before plain_b, preserved).
-    assert_eq!(peers, vec![anchor_a, anchor_b, plain_a, plain_b]);
-}
-
-#[test]
-fn partition_peer_with_one_anchor_identity_among_many_qualifies() {
-    // A peer can host multiple context identities — partition matches
-    // if ANY identity intersects the anchor set.
-    let peer = dummy_peer(1);
-    let pk_admin = dummy_pk(0xAA);
-    let pk_other_context = dummy_pk(0xBB);
-
-    let node_state = node_state_with_peer_identities([(
-        peer,
-        [pk_admin, pk_other_context].into_iter().collect(),
-    )]);
-
-    let anchors: BTreeSet<PublicKey> = [pk_admin].into_iter().collect();
-
-    let mut peers = vec![peer];
-    let count = partition_peers_anchor_first(&mut peers, &node_state, &anchors);
-    assert_eq!(count, 1);
-}
-
-#[test]
-fn partition_peer_with_only_non_anchor_identities_does_not_qualify() {
-    let peer = dummy_peer(1);
-    let pk_member = dummy_pk(0xCC);
-    let pk_admin = dummy_pk(0xAA);
-
-    let node_state = node_state_with_peer_identities([(peer, [pk_member].into_iter().collect())]);
-
-    let anchors: BTreeSet<PublicKey> = [pk_admin].into_iter().collect();
-
-    let mut peers = vec![peer];
-    let count = partition_peers_anchor_first(&mut peers, &node_state, &anchors);
-    assert_eq!(count, 0);
-}
-
-/// `partition_peers_anchor_first` works against any
-/// `SyncStateAccess` impl, not just `NodeState`. This test exercises
-/// the same partition behaviour using the `MockSyncStateAccess`
-/// fixture — a pure unit test surface for the partition helper that
-/// could be extended to higher-level sync code as more dependencies
-/// become mockable.
-///
-/// Pinning this demonstrates the test surface promised in the
-/// trait's doc: `sync/` code that goes through `SyncStateAccess`
-/// can be exercised without a `NodeState` at all.
-#[test]
-fn partition_works_against_mock_sync_state_access() {
-    use crate::sync::state_access_mock::{MockSyncStateAccess, SyncStateAccessCall};
-
-    let anchor_peer = dummy_peer(1);
-    let plain_peer = dummy_peer(2);
-    let pk_admin = dummy_pk(0xAA);
-
-    let mock = MockSyncStateAccess::default();
-    mock.insert_peer_identities(anchor_peer, [pk_admin].into_iter().collect());
-
-    let anchors: BTreeSet<PublicKey> = [pk_admin].into_iter().collect();
-
-    let mut peers = vec![plain_peer, anchor_peer];
-    let count = partition_peers_anchor_first(&mut peers, &mock, &anchors);
-    assert_eq!(count, 1, "exactly one anchor in the input");
-    assert_eq!(
-        peers,
-        vec![anchor_peer, plain_peer],
-        "anchor should come first"
-    );
-
-    // Both peer-identity lookups should have been observed in order.
-    let calls = mock.calls();
-    assert_eq!(
-        calls,
-        vec![
-            SyncStateAccessCall::PeerIdentities(plain_peer),
-            SyncStateAccessCall::PeerIdentities(anchor_peer),
-        ],
-        "partition should query peer_identities exactly once per input peer, in order"
-    );
-}
-
-// =========================================================================
 // `reconcile_cooldown` / `record_reconcile_*` — backoff for the
 // reconcile-after-divergence path
 // =========================================================================
@@ -545,6 +349,7 @@ fn partition_works_against_mock_sync_state_access() {
 use std::time::Duration;
 
 use calimero_primitives::context::ContextId;
+use dashmap::DashMap;
 
 use super::{
     reconcile_cooldown, reconcile_remaining_cooldown, record_reconcile_failure,
